user/requirements.txt
Flask>=2.2.2
psycopg2-binary==2.8.6
opentracing
jaeger-client
